Clean up and reusablity
