+++ elide-async/src/main/java/com/yahoo/elide/async/hooks/ExecuteQueryHook.java
@@ -7,9 +7,12 @@ package com.yahoo.elide.async.hooks;
 
 import com.yahoo.elide.annotation.LifeCycleHookBinding;
 import com.yahoo.elide.async.models.AsyncQuery;
-import com.yahoo.elide.async.models.QueryStatus;
+import com.yahoo.elide.async.models.QueryType;
 import com.yahoo.elide.async.service.AsyncExecutorService;
+import com.yahoo.elide.async.service.AsyncQueryThread;
+import com.yahoo.elide.core.exceptions.InvalidOperationException;
 import com.yahoo.elide.functions.LifeCycleHook;
+import com.yahoo.elide.graphql.QueryRunner;
 import com.yahoo.elide.security.ChangeSpec;
 import com.yahoo.elide.security.RequestScope;
 
@@ -18,20 +21,27 @@ import java.util.Optional;
 /**
  * LifeCycle Hook for execution of AsyncQuery.
  */
-public class ExecuteQueryHook implements LifeCycleHook<AsyncQuery> {
-
-    private AsyncExecutorService asyncExecutorService;
+public class ExecuteQueryHook extends AsyncHook implements LifeCycleHook<AsyncQuery> {
 
     public ExecuteQueryHook (AsyncExecutorService asyncExecutorService) {
-        this.asyncExecutorService = asyncExecutorService;
+        super(asyncExecutorService);
     }
 
     @Override
     public void execute(LifeCycleHookBinding.Operation operation, LifeCycleHookBinding.TransactionPhase phase,
                         AsyncQuery query, RequestScope requestScope, Optional<ChangeSpec> changes) {
-        validateOptions(query);
-        if (query.getStatus() == QueryStatus.QUEUED && query.getResult() == null) {
-            asyncExecutorService.executeQuery(query, requestScope.getUser(), requestScope.getApiVersion());
+        validateOptions(query, requestScope);
+        AsyncQueryThread queryWorker = new AsyncQueryThread(query, requestScope.getUser(), getAsyncExecutorService(),
+                requestScope.getApiVersion());
+        executeAsync(query, queryWorker);
+    }
+
+    protected void validateOptions(AsyncQuery query, RequestScope requestScope) {
+        if (query.getQueryType().equals(QueryType.GRAPHQL_V1_0)) {
+            QueryRunner runner = getAsyncExecutorService().getRunners().get(requestScope.getApiVersion());
+            if (runner == null) {
+                throw new InvalidOperationException("Invalid API Version");
+            }
         }
     }
 }
