+++ elide-async/src/main/java/com/yahoo/elide/async/models/AsyncQuery.java
@@ -5,32 +5,15 @@
  */
 package com.yahoo.elide.async.models;
 
-import com.yahoo.elide.Elide;
-import com.yahoo.elide.ElideResponse;
 import com.yahoo.elide.annotation.DeletePermission;
 import com.yahoo.elide.annotation.Include;
 import com.yahoo.elide.annotation.ReadPermission;
 import com.yahoo.elide.annotation.UpdatePermission;
-import com.yahoo.elide.async.service.AsyncExecutorService;
-import com.yahoo.elide.graphql.QueryRunner;
-import com.yahoo.elide.security.User;
-
-import org.apache.http.NameValuePair;
-import org.apache.http.NoHttpResponseException;
-import org.apache.http.client.utils.URIBuilder;
 
 import lombok.Data;
-import lombok.extern.slf4j.Slf4j;
-
-import java.net.URISyntaxException;
-import java.util.Date;
-import java.util.Map;
-import java.util.UUID;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.ws.rs.core.MultivaluedHashMap;
-import javax.ws.rs.core.MultivaluedMap;
 
 /**
  * Model for Async Query.
@@ -43,7 +26,6 @@ import javax.ws.rs.core.MultivaluedMap;
 @UpdatePermission(expression = "Prefab.Role.None")
 @DeletePermission(expression = "Prefab.Role.None")
 @Data
-@Slf4j
 public class AsyncQuery extends AsyncAPI {
     @Embedded
     private AsyncQueryResult result;
@@ -51,81 +33,5 @@ public class AsyncQuery extends AsyncAPI {
     @Override
     public void setResult(AsyncAPIResult result) {
         this.result = (AsyncQueryResult) result;
-    }
-
-    @Override
-    public AsyncAPIResult executeRequest(AsyncExecutorService service, User user, String apiVersion)
-            throws URISyntaxException, NoHttpResponseException {
-        ElideResponse response = null;
-        log.debug("AsyncQuery Object from request: {}", this);
-        if (queryType.equals(QueryType.JSONAPI_V1_0)) {
-            response = executeJsonApiRequest(service.getElide(), user, apiVersion);
-        } else if (queryType.equals(QueryType.GRAPHQL_V1_0)) {
-            response = executeGraphqlRequest(service.getRunners(), user, apiVersion);
-        }
-        nullResponseCheck(response);
-
-        AsyncQueryResult queryResult = new AsyncQueryResult();
-        queryResult.setHttpStatus(response.getResponseCode());
-        queryResult.setCompletedOn(new Date());
-        queryResult.setResponseBody(response.getBody());
-        queryResult.setContentLength(response.getBody().length());
-        //TODO Add recordcount to queryResultObj
-        return queryResult;
-    }
-
-    /**
-     * This method parses the url and gets the query params.
-     * And adds them into a MultivaluedMap to be used by underlying Elide.get method
-     * @param uri URIBuilder instance
-     * @return MultivaluedMap with query parameters
-     */
-    private MultivaluedMap<String, String> getQueryParams(URIBuilder uri) {
-        MultivaluedMap<String, String> queryParams = new MultivaluedHashMap<String, String>();
-        for (NameValuePair queryParam : uri.getQueryParams()) {
-            queryParams.add(queryParam.getName(), queryParam.getValue());
-        }
-        return queryParams;
-    }
-
-    /**
-     * This method parses the url and gets the query params.
-     * And retrieves path to be used by underlying Elide.get method
-     * @param uri URIBuilder instance
-     * @throws URISyntaxException URISyntaxException from malformed or incorrect URI
-     * @return Path extracted from URI
-     */
-    private String getPath(URIBuilder uri) {
-        return uri.getPath();
-    }
-
-    private ElideResponse executeJsonApiRequest(Elide elide, User user, String apiVersion) throws URISyntaxException {
-        UUID requestUUId = UUID.fromString(requestId);
-        URIBuilder uri = new URIBuilder(query);
-        MultivaluedMap<String, String> queryParams = getQueryParams(uri);
-        log.debug("Extracted QueryParams from AsyncQuery Object: {}", queryParams);
-
-        //TODO - we need to add the baseUrlEndpoint to the queryObject.
-        ElideResponse response = elide.get("", getPath(uri), queryParams, user, apiVersion, requestUUId);
-        log.debug("JSONAPI_V1_0 getResponseCode: {}, JSONAPI_V1_0 getBody: {}",
-                response.getResponseCode(), response.getBody());
-        return response;
-    }
-
-    private ElideResponse executeGraphqlRequest(Map<String, QueryRunner> runners, User user, String apiVersion)
-            throws URISyntaxException {
-        UUID requestUUId = UUID.fromString(requestId);
-        QueryRunner runner = runners.get(apiVersion);
-        //TODO - we need to add the baseUrlEndpoint to the queryObject.
-        ElideResponse response = runner.run("", query, user, requestUUId);
-        log.debug("GRAPHQL_V1_0 getResponseCode: {}, GRAPHQL_V1_0 getBody: {}",
-                response.getResponseCode(), response.getBody());
-        return response;
-    }
-
-    private void nullResponseCheck(ElideResponse response) throws NoHttpResponseException {
-        if (response == null) {
-            throw new NoHttpResponseException("Response for request returned as null");
-        }
     }
 }
