+++ elide-async/src/main/java/com/yahoo/elide/async/service/AsyncAPIThread.java
@@ -5,14 +5,11 @@
  */
 package com.yahoo.elide.async.service;
 
-import com.yahoo.elide.async.models.AsyncAPI;
 import com.yahoo.elide.async.models.AsyncAPIResult;
-import com.yahoo.elide.security.User;
 
 import org.apache.http.NoHttpResponseException;
 
 import lombok.Data;
-import lombok.extern.slf4j.Slf4j;
 
 import java.net.URISyntaxException;
 import java.util.concurrent.Callable;
@@ -22,25 +19,12 @@ import java.util.concurrent.Callable;
  * It will also update the query status and result object at different
  * stages of execution.
  */
-@Slf4j
 @Data
-public class AsyncAPIThread implements Callable<AsyncAPIResult> {
-
-    private AsyncAPI queryObj;
-    private User user;
-    private AsyncExecutorService service;
-    private String apiVersion;
+public abstract class AsyncAPIThread implements Callable<AsyncAPIResult> {
 
     @Override
     public AsyncAPIResult call() throws NoHttpResponseException, URISyntaxException {
          return processQuery();
-    }
-
-    public AsyncAPIThread(AsyncAPI queryObj, User user, AsyncExecutorService service, String apiVersion) {
-        this.queryObj = queryObj;
-        this.user = user;
-        this.service = service;
-        this.apiVersion = apiVersion;
     }
 
    /**
@@ -50,10 +34,5 @@ public class AsyncAPIThread implements Callable<AsyncAPIResult> {
     * @throws URISyntaxException URISyntaxException
     * @throws NoHttpResponseException NoHttpResponseException
     */
-    protected AsyncAPIResult processQuery() throws URISyntaxException, NoHttpResponseException {
-        // Create AsyncQueryResultBase Object
-        AsyncAPIResult queryResultBaseObj = queryObj.executeRequest(service, user, apiVersion);
-
-        return queryResultBaseObj;
-    }
+    protected abstract AsyncAPIResult processQuery() throws URISyntaxException, NoHttpResponseException;
 }
