+++ elide-async/src/test/java/com/yahoo/elide/async/service/AsyncExecutorServiceTest.java
@@ -13,9 +13,9 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-
 import com.yahoo.elide.Elide;
 import com.yahoo.elide.ElideSettingsBuilder;
+import com.yahoo.elide.async.models.AsyncAPIResult;
 import com.yahoo.elide.async.models.AsyncQuery;
 import com.yahoo.elide.async.models.QueryStatus;
 import com.yahoo.elide.async.models.QueryType;
@@ -29,11 +29,9 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 
-import java.net.URISyntaxException;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.TimeoutException;
+import java.util.concurrent.Callable;
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class AsyncExecutorServiceTest {
@@ -74,12 +72,14 @@ public class AsyncExecutorServiceTest {
 
     //Test for executor hook execution
     @Test
-    public void testExecuteQueryFail() throws ExecutionException, TimeoutException, InterruptedException,
-           NoHttpResponseException, URISyntaxException {
+    public void testExecuteQueryFail() throws Exception {
        AsyncQuery queryObj = mock(AsyncQuery.class);
        when(queryObj.getAsyncAfterSeconds()).thenReturn(10);
-       when(queryObj.executeRequest(service, testUser, NO_VERSION)).thenThrow(new NoHttpResponseException(null));
-       service.executeQuery(queryObj, testUser, NO_VERSION);
+
+       Callable<AsyncAPIResult> mockCallable = mock(Callable.class);
+       when(mockCallable.call()).thenThrow(new NoHttpResponseException(null));
+
+       service.executeQuery(queryObj, mockCallable);
        verify(queryObj, times(1)).setStatus(QueryStatus.PROCESSING);
        verify(queryObj, times(1)).setStatus(QueryStatus.FAILURE);
     }
@@ -96,7 +96,8 @@ public class AsyncExecutorServiceTest {
         when(queryObj.getRequestId()).thenReturn(id);
         when(queryObj.getQueryType()).thenReturn(QueryType.JSONAPI_V1_0);
         when(queryObj.getAsyncAfterSeconds()).thenReturn(10);
-        service.executeQuery(queryObj, testUser, NO_VERSION);
+        AsyncAPIThread queryThread = new AsyncQueryThread(queryObj, testUser, service, NO_VERSION);
+        service.executeQuery(queryObj, queryThread);
         verify(queryObj, times(1)).setStatus(QueryStatus.PROCESSING);
         verify(queryObj, times(1)).setStatus(QueryStatus.COMPLETE);
     }
@@ -108,7 +109,8 @@ public class AsyncExecutorServiceTest {
         AsyncQuery queryObj = mock(AsyncQuery.class);
         when(queryObj.getAsyncAfterSeconds()).thenReturn(0);
         when(queryObj.getQueryUpdateWorker()).thenReturn(asyncAPIUpdateThread);
-        service.executeQuery(queryObj, testUser, NO_VERSION);
+        AsyncAPIThread queryThread = new AsyncQueryThread(queryObj, testUser, service, NO_VERSION);
+        service.executeQuery(queryObj, queryThread);
         service.completeQuery(queryObj, testUser, NO_VERSION);
         verify(queryObj, times(1)).setStatus(QueryStatus.PROCESSING);
         verify(queryObj, times(2)).getQueryUpdateWorker();
